pyproject.toml
[project]
name = "replayt-opentelemetry-exporter"
version = "0.1.0"
description = "OpenTelemetry traces and metrics for replayt workflow runs: Python project using replayt; see docs for scope."
requires-python = ">=3.11"
dependencies = [
  "opentelemetry-api>=1.20.0",
  "opentelemetry-sdk>=1.20.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6.0",
]
otlp = [
  "opentelemetry-exporter-otlp-proto-http>=1.20.0",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
